+++ wsidicomizer/sources/openslide_like/openslide_like_image_data.py
@@ -56,7 +56,7 @@ class OpenSlideLikeImageData(PixelImageData):
 
     @property
     def optical_paths(self) -> list[str]:
-        return ["0"]
+        return ["1"]
 
     @property
     def blank_color(self) -> int | tuple[int, int, int]: