+++ wsidicomizer/sources/opentile/opentile_metadata.py
@@ -61,7 +61,7 @@ class OpenTileMetadata(WsiDicomizerMetadata):
         )
         if objectives is not None or icc_profile is not None:
             optical_paths = [
-                OpticalPath("0", objective=objectives, icc_profile=icc_profile)
+                OpticalPath("1", objective=objectives, icc_profile=icc_profile)
             ]
         else:
             optical_paths = []