+++ CHANGELOG.md
@@ -217,7 +217,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Support for pydicom 3.0. This requires Pyhton >3.10.
+- Support for pydicom 3.0. This requires Python >3.10.
 
 ### Removed
 
@@ -282,7 +282,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Replaced `Dataset` based metadata by `modules`-parameter to `open()` and `convert()` with metadata models from `WsiDicom`. Use the `metadata`-parameter to define metadata that should override any metadata found in the source file, and the`default_metadata`-parameter to define metadata that should be used if no other metadata is defiend.
+- Replaced `Dataset` based metadata by `modules`-parameter to `open()` and `convert()` with metadata models from `WsiDicom`. Use the `metadata`-parameter to define metadata that should override any metadata found in the source file, and the`default_metadata`-parameter to define metadata that should be used if no other metadata is defined.
 
 ### Fixed